Enhance behavior cloning and teacher functionality
- Added demonstration fingerprinting to save and load encoded demonstrations, allowing for better tracking of teacher flight settings and configurations.
- Improved the handling of visual loss cues in the privileged velocity teacher, enabling it to take cues from geometry and adjust descent behavior accordingly.
- Updated the reward design logic to ensure that no two workers can fly the same UAV/UGV pair simultaneously, preventing conflicts and ensuring smoother operation.
- Introduced new tests to validate the behavior of the teacher under various visibility conditions and to ensure the integrity of the reward design mechanism.
- Refactored the teacher's descent logic to be more configurable and responsive to visual cues, enhancing overall flight performance.

+++ Ontology_RGAT_UAV_RL_ISAAC_PX4/python/ontology_rgat/perception/keypoint_encoder.py
@@ -139,7 +139,9 @@ class ShinKeypointEncoder(nn.Module):
         grid_y, grid_x = torch.meshgrid(ys, xs, indexing="ij")
         x = (probabilities * grid_x.reshape(1, 1, -1)).sum(-1)
         y = (probabilities * grid_y.reshape(1, 1, -1)).sum(-1)
-        return torch.stack((x, y), dim=-1).to(heatmaps.dtype)
+        # Kept in fp32: under bfloat16 autocast the heatmaps' dtype would
+        # quantise a coordinate in [-1, 1] to ~2 px steps.
+        return torch.stack((x, y), dim=-1)
 
     def forward(self, image: torch.Tensor) -> KeypointEncoderOutput:
         if image.ndim != 4 or image.shape[1] != 1:
